Aligns JDBC behavior with upstream
This corrects some small differences in environment variables and
defaults vs upstream. More importantly, it fixes a bug where limit
applied to spans as opposed to traces.
